Integrate a new option in `launch` command (#15)
For each pintool, the user can add in the config file a new parameter named
alt_prev_step.
Scat `launch` command can take `--alt_prev` as option to define the prev_step
according to the field alt_prev_step in the config file
When `coupleres` is the previous step, the python script adds the right
comand line argument for the pintool to use.